+++ src/General/Modules/SetupAndMenus/Header/QELogImport.js
@@ -2,10 +2,10 @@ import React, { useEffect } from "react";
 import axios from "axios";
 import { useTranslation } from "react-i18next";
 import { Button, Dialog, DialogActions, DialogContent, DialogTitle, Grid, InputLabel, FormControl, Select, MenuItem, Typography, Tooltip } from "@material-ui/core";
-import LogLinkInput from "../../LogImport/LogLinkInput";
+import LogLinkInput from "../../../SystemTools/LogImport/LogLinkInput";
 import { warcraftLogReportID, logDifficulty, importSummaryData, importDamageLogData } from "../../CooldownPlanner/Functions/Functions";
 import { classColoursJS } from "../../CooldownPlanner/Functions/ClassColourFunctions";
-import FightSelectorButton from "../../LogImport/FightSelectorButton";
+import FightSelectorButton from "../../../SystemTools/LogImport/FightSelectorButton";
 import bossIcons from "../../CooldownPlanner/Functions/IconFunctions/BossIcons";
 import { convertLogSpellOutput, convertLogStatOutput } from "../../Player/PlayerUtilities";
